apps/api/pyproject.toml
[project]
name = "vendorcheck-api"
version = "0.2.0"
description = "VendorCheck API — AWS-first multi-tenant backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "pydantic-settings>=2.0",
    "python-multipart>=0.0.9",
    # Database
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg>=0.29",
    "psycopg2-binary>=2.9",
    # Auth (Cognito JWKS)
    "python-jose[cryptography]>=3.3",
    "httpx>=0.27",
    # AWS SDK
    "boto3>=1.34",
    # Legacy (kept during migration)
    "supabase>=2.0",
    "PyJWT>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "aiosqlite>=0.20",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
